+++ tests/e2e/game/full-game.test.js
@@ -90,6 +90,7 @@ describe("B - Full game of 31 players with all roles", () => {
                 expect(game.options.roles.areRevealedOnDeath).to.be.true;
                 expect(game.options.roles.sheriff.electedAt.turn).to.be.equal(1);
                 expect(game.options.roles.sheriff.electedAt.phase).to.be.equal("night");
+                expect(game.options.roles.bigBadWolf.isPowerlessIfWerewolfDies).to.be.true;
                 expect(game.options.roles.seer.canSeeRoles).to.be.true;
                 expect(game.options.roles.guard.canProtectTwice).to.be.false;
                 expect(game.options.roles.fox.isPowerlessIfMissesWerewolf).to.be.true;